MAINT: Refactor OCG functions out from writer
